lambda/thumbnails/requirements.txt
boto3==1.7.72
botocore==1.10.72
docutils==0.14
jmespath==0.9.3
Pillow==5.2.0
python-dateutil==2.7.3
s3transfer==0.1.13
six==1.11.0
